Replace Windows installer FilesInUse dialog text
Windows installers must provide a FilesInUse dialog, with the intent
of minimizing the need to restart the system after software
installations.  For reasons related to the LSA cache, the KfW
installer always schedules a post-installation reboot (see commit
50a3c3cbeab32577fba2b21deb72a64015c48ec7).  Reword the FilesInUse
dialog text to recommend clicking Ignore, causing the conflicting
files to replaced on the reboot.  (The affected files are likely C
runtime DLLs, and the existing versions will almost certainly suffice
if KfW programs are run prior to a reboot.)

ticket: 9135 (new)