Add Message-Authenticator support to libkrad
draft-ietf-radext-deprecating-radius-03 describes multiple mitigation
measures for recent vulnerabilities discovered in the RADIUS protocol,
the BlastRADIUS (CVE-2024-3596).  One of these measures is enforcement
of the Message-Authenticator attribute.

Message-Authenticator was initially defined in RFC2869 (2000) as an
extension to improve authenticity and integrity verification.  The
BlastRADIUS vulnerability is now pushing vendors to generalize the use
of this attibute.  FreeRADIUS enforces Message-Authenticator as of
versions 3.2.5 and 3.0.27.  libkrad has to support
Message-Authenticator in order to remain compatible with these
implementations.

[ghudson@mit.edu: adjusted style and naming; simplified some
functions; edited commit message]