Refactor libapputils handling of socket addresses
Add private libkrb5 APIs for sockaddr-to-krb5-address conversion and
for printing socket addresses.  Use them in the KDC, kadmind,
kprop/kpropd, and trace logging.

In net-server.c, do not convert the local and remote socket addresses
to krb5_address; instead pass them directly to dispatch().  Convert
the addresses where needed in the KDC and kadmind.