Add support for systemd socket activation
If LISTEN_PID and LISTEN_FDS are set in the environment, expect
listener sockets to be present at fds starting at 3.  If any match a
configured listener address and port, use it instead of creating a new
socket.

[ghudson@mit.edu: combined two commits; changed sa_compare() to
sa_equal(); restructured changes to fetch listener sockets into a list
once and to match on socket type as well as address; added tests]

ticket: 9157 (new)