Build PKINIT on Windows
Include PKINIT in the Windows build if OPENSSL_DIR is set.

In the installer, require PKINIT to be built, and require
OPENSSL_VERSION to be set in the environment.  Include the PKINIT
module and a copy of the OpenSSL libcrypto DLL.

Document the OpenSSL dependency in windows/README.  Also clarify the
Visual Studio dependencies, and restore the prep-windows build step
which was accidentally removed by commit
5f33fb2cf19562175c8271cb9c54a67212f63b93.

[ghudson@mit.edu: added CI and installer integration; edited README
and commit message]

ticket: 9162 (new)