Fix getsockname() call in Windows localaddr
When querying local IP addresses, local_addr_fallback_kludge() is
called on Windows if gethostname() or gethostbyname() fails.  This
function's call to getsockname() erroneously passes a small integer as
the third argument instead of a pointer, likely causing a process
crash.  This bug has been present since 1.1, but is rarely encountered
due to the rarity of requests for addressful tickets.  Fix it.

[ghudson@mit.edu: rewrote commit message]

ticket: 9171 (new)