Simplify and fix libkdb_ldap validation checks
In libkdb_ldap, where we validate arguments with null checks (which is
done inconsistently, mostly in policy-related functions), do it as
early as possible and return EINVAL without using the cleanup handler.
Remove code to set (generally misleading) extended error messages, as
we don't expect these conditions to be true in practice.

Fixes a potential crash if krb5_ldap_read_policy() is called with a
null policy argument.  (It never is, so this isn't a user-visible
bug.)

[ghudson@mit.edu: added additional changes; rewrote commit message]