Refactor dnsglue functions using k5input
Use k5input to make the dnsglue.c and dnssrv.c functions simpler and
more robust.  Rename the dnsglue functions to use the now-preferred
k5_ prefix instead of krb5int_.  Use stdint type names instead of
unsigned char and unsigned short.  Remove some out-of-date comments.

Change the query loop in k5_dns_init() to continue expanding the
buffer if the reported answer length is exactly equal to the buffer
size, as that is the observed behavior from the glibc res_nquery()
when the buffer size is too small for the answer.