Avoid undefined memcpy in asn1_encode.c
The C standard specifies that passing null pointers to most standard
library functions results in undefined behavior (C99 7.1.4).  This
applies to memcpy() even when the length is 0.  insert_bytes() in
asn1_encode.c may be called with a null pointer from an empty
krb5_data or other counted value in a structure to be encoded.  Do not
call memcpy() in this case.

Reported by Kirill Furman.

ticket: 9175